requirements.txt
coverage==7.13.5
isort==8.0.1
ruff==0.15.8
mypy==1.20.0

# Interactive shell dependencies
rich>=13.0.0
PyYAML>=6.0
